+++ backend/main.go
@@ -60,8 +60,6 @@ func main() {
 	// r.POST("/dogs", dogs.CreateDog)
 	// r.PUT("/dogs/:id", dogs.UpdateDog)
 	// r.DELETE("/dogs/:id", dogs.DeleteDog)
-	r.PUT("/dogs/:id", dog.UpdateDog)
-	r.DELETE("/dogs/:id", dog.DeleteDog)
 	r.POST("/sponsorships/one-time", sponsorship.CreateOneTimeSponsorship)
 	r.GET("/genders", gender.GetAll)
 	r.GET("/vaccines", vaccine.GetAll)
@@ -114,6 +112,8 @@ func main() {
 	r.DELETE("/volunteer/:id", volunteers.DeleteVolunteer)
 	r.GET("/skills", volunteers.GetAllSkills)     //new
 	r.GET("/statusfv", volunteers.GetAllStatusFV) //new
+	r.PUT("/volunteer/:id/status", volunteers.UpdateVolunteerStatus)
+	
 	// 7) Routes (protected)
 	protected := r.Group("/")
 	protected.Use(middlewares.Authorizes())
