+++ frontend/src/pages/dashboard/Volunteer/style.css
@@ -1,0 +1,240 @@
+/* Updated styles to match the reference design */
+
+.admin-approvals .admin-toolbar {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 16px;
+  margin-bottom: 24px;
+  padding: 16px 0;
+}
+
+.admin-approvals .filter-group {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+}
+
+.admin-approvals .filter-group label {
+  font-weight: 500;
+  color: #374151;
+}
+
+.admin-approvals .filter-group select {
+  padding: 8px 12px;
+  border: 1px solid #d1d5db;
+  border-radius: 8px;
+  background: white;
+  font-size: 14px;
+}
+
+.admin-approvals .submit-btn {
+  background: #3b82f6;
+  color: white;
+  border: none;
+  padding: 8px 16px;
+  border-radius: 8px;
+  font-weight: 500;
+  cursor: pointer;
+}
+
+.admin-approvals .submit-btn:hover {
+  background: #2563eb;
+}
+
+.admin-approvals .submit-btn:disabled {
+  opacity: 0.6;
+  cursor: not-allowed;
+}
+
+.admin-approvals .loading,
+.admin-approvals .empty {
+  padding: 32px;
+  text-align: center;
+  color: #6b7280;
+  background: #f9fafb;
+  border: 1px solid #e5e7eb;
+  border-radius: 12px;
+  font-size: 16px;
+}
+
+.admin-approvals .error-message {
+  background: #fef2f2;
+  color: #dc2626;
+  padding: 12px 16px;
+  border: 1px solid #fecaca;
+  border-radius: 8px;
+  margin-bottom: 16px;
+}
+
+.admin-approvals .cards {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(380px, 1fr));
+  gap: 20px;
+}
+
+.admin-approvals .card {
+  background: white;
+  border: 1px solid #e5e7eb;
+  border-radius: 16px;
+  padding: 20px;
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1), 0 1px 2px rgba(0, 0, 0, 0.06);
+  transition: box-shadow 0.2s ease-in-out;
+}
+
+.admin-approvals .card:hover {
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1), 0 2px 4px rgba(0, 0, 0, 0.06);
+}
+
+.admin-approvals .card-row {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+  padding: 8px 0;
+  border-bottom: 1px solid #f3f4f6;
+}
+
+.admin-approvals .card-row:last-of-type {
+  border-bottom: none;
+}
+
+.admin-approvals .card-row.head {
+  padding-bottom: 16px;
+  margin-bottom: 16px;
+  border-bottom: 2px solid #e5e7eb;
+  align-items: center;
+}
+
+.admin-approvals .name {
+  font-weight: 600;
+  font-size: 18px;
+  color: #111827;
+}
+
+.admin-approvals .label {
+  color: #6b7280;
+  font-size: 14px;
+  font-weight: 500;
+  min-width: 120px;
+  flex-shrink: 0;
+}
+
+.admin-approvals .value {
+  color: #374151;
+  font-size: 14px;
+  text-align: right;
+  flex: 1;
+}
+
+.admin-approvals .badge {
+  padding: 6px 12px;
+  border-radius: 20px;
+  font-size: 12px;
+  font-weight: 600;
+  text-transform: uppercase;
+  letter-spacing: 0.5px;
+  border: 1px solid transparent;
+}
+
+.admin-approvals .badge.pending { 
+  background: #dbeafe; 
+  color: #1d4ed8; 
+  border-color: #93c5fd; 
+}
+
+.admin-approvals .badge.approved { 
+  background: #d1fae5; 
+  color: #059669; 
+  border-color: #6ee7b7; 
+}
+
+.admin-approvals .badge.rejected { 
+  background: #fee2e2; 
+  color: #dc2626; 
+  border-color: #fca5a5; 
+}
+
+.admin-approvals .badge.unknown,
+.admin-approvals .badge.none { 
+  background: #f3f4f6; 
+  color: #6b7280; 
+  border-color: #e5e7eb; 
+}
+
+.admin-approvals .actions {
+  display: flex;
+  justify-content: center;
+  gap: 12px;
+  margin-top: 20px;
+  padding-top: 16px;
+  border-top: 1px solid #f3f4f6;
+}
+
+.admin-approvals .btn {
+  border: none;
+  border-radius: 8px;
+  padding: 10px 20px;
+  cursor: pointer;
+  font-weight: 600;
+  font-size: 14px;
+  transition: all 0.2s ease-in-out;
+  min-width: 100px;
+}
+
+.admin-approvals .btn.approve { 
+  background: #10b981; 
+  color: white; 
+}
+
+.admin-approvals .btn.approve:hover:not(:disabled) { 
+  background: #059669; 
+  transform: translateY(-1px);
+}
+
+.admin-approvals .btn.reject  { 
+  background: #ef4444; 
+  color: white; 
+}
+
+.admin-approvals .btn.reject:hover:not(:disabled) { 
+  background: #dc2626; 
+  transform: translateY(-1px);
+}
+
+.admin-approvals .btn:disabled { 
+  opacity: 0.5; 
+  cursor: not-allowed;
+  transform: none;
+}
+
+/* Responsive adjustments */
+@media (max-width: 768px) {
+  .admin-approvals .cards {
+    grid-template-columns: 1fr;
+    gap: 16px;
+  }
+  
+  .admin-approvals .admin-toolbar {
+    flex-direction: column;
+    align-items: stretch;
+    gap: 12px;
+  }
+  
+  .admin-approvals .filter-group {
+    justify-content: space-between;
+  }
+  
+  .admin-approvals .card-row {
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 4px;
+  }
+  
+  .admin-approvals .value {
+    text-align: left;
+  }
+  
+  .admin-approvals .actions {
+    flex-direction: column;
+  }
+}
