--- frontend/src/pages/public/ZoneCageManagement/style.css
@@ -1,331 +1,0 @@
-.zone-cage-management-container {
-  background-image: url('../../../assets/donate_background.webp');
-  background-size: cover;
-  background-position: center;
-  background-repeat: no-repeat;
-  min-height: 75vh;
-  display: flex;
-  justify-content: center; /* Changed from 'left' to 'space-between' */
-  align-items: center;
-  color: rgb(0, 0, 0);
-  padding: 140px;
-  position: relative;
-  font-family: 'Anakotmai', sans-serif;
-  gap: 400px; /* Add gap between the two sections */
-} 
-
-.zone-cage-management-container::before {
-  content: '';
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: #171C2D;
-  opacity: 0.8;
-  z-index: 1;
-}
-
-.zone-cage-management-form {
-  position: relative;
-  z-index: 2;
-  max-width: 700px;
-  max-height: 300px;
-  background-color: rgba(255, 255, 255, 0.9);
-  padding: 20px;
-  border-radius: 10px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  text-align: center; /* Center all content inside */
-}
-
-.zone-cage-management-form h1 {
-  text-align: center;
-  color: #1e40af;
-  font-size: 28px;
-  font-weight: 600;
-  margin-bottom: 30px;
-}
-
-/* Remove the conflicting form-group rule and keep only this one */
-.form-group {
-  display: flex;
-  align-items: center;
-  justify-content: center; /* Center the form group content */
-  gap: 12px;
-  margin-bottom: 20px; /* Add some spacing between form groups */
-}
-
-.form-group label {
-  margin: 0;
-  text-align: right;
-  font-size: 20px;
-  min-width: 80px; /* Give labels consistent width */
-}
-
-.form-group select {
-  width: 240px;
-  padding: 8px; /* Add padding for better appearance */
-  color: #000;
-  background-color: #fff; /* Change to white for better readability */
-  border: 1px solid #ccc;
-  border-radius: 10px;
-}
-
-.zone-cage-management-form button {
-  background-color: #d97706;
-  color: white;
-  padding: 10px 20px;
-  border: solid 1px #d97706;
-  border-radius: 5px;
-  cursor: pointer;
-  font-size: 14px;
-  margin: 5px;
-}
-/* Remove the form-container class as it's causing conflicts */
-.zone-cage-management-form button:hover {
-  background-color: #f59e0b;
-  color: white;
-}
-
-
-/**/
-/**/
-/* Information Zone */
-.information-zone-container {
-  position: relative;
-  background-color: rgba(255, 255, 255, 0.8);
-  padding: 12px;
-  border-radius: 8px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  margin-top: 20px;
-  z-index: 2;
-  min-height: 500px;
-  min-width: 500px;
-  height: auto;
-  display: flex;
-  flex-direction: column;
-}
-
-.information-form {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  justify-content: center;   /* Add this */
-  align-items: center; 
-}
-
-.information-form h2 {
-  text-align: center;
-  color: #1e40af;
-  font-size: 20px;
-  font-weight: 600;
-  margin-bottom: 12px;
-  flex-shrink: 0; /* Prevent the title from shrinking */
-}
-
-.loader-container {
-  flex: 1;
-  display: flex;
-  justify-content: center;
-  align-items: center; /* Offset the title height to true center */
-  flex-direction: column;
-  margin-top: 170px;
-  gap: 10px;
-}
-
-/* HTML: <div class="loader"></div> */
-.loader {
-  width: 40px;
-  aspect-ratio: 1;
-  outline: 2px solid #fff;
-  background:
-    radial-gradient(farthest-side,#000 90%,#0000) 0 0/8px 8px no-repeat,
-    conic-gradient(from  90deg at 10px 10px,#0000 90deg,#fff 0),
-    conic-gradient(from -90deg at 30px 30px,#0000 90deg,#fff 0);
-  animation: l3 1s infinite;
-}
-@keyframes l3 {
-  0%   {background-position:0    0   ;clip-path:circle(15px at left  4px top    4px)}
-  25%  {background-position:100% 0   ;clip-path:circle(15px at right 4px top    4px)}
-  50%  {background-position:100% 100%;clip-path:circle(15px at right 4px bottom 4px)}
-  75%  {background-position:0    100%;clip-path:circle(15px at left  4px bottom 4px)}
-  100% {background-position:0    0   ;clip-path:circle(15px at left  4px top    4px)}
-}
-
-.information-form button {
-  background-color: #d97706;
-  color: white;
-  border: solid 1px #d97706;
-  border-radius: 5px;
-  padding: 10px 20px;
-  cursor: pointer;
-  font-size: 14px;
-  margin: 5px;
-  flex-direction: column-reverse;
-}
-
-.information-form button:hover {
-  background-color: #f59e0b;
-  color: white;
-}
-
-
-.info-box {
-  position: relative;
-  background-color: rgba(255, 255, 255, 0.9);
-  border: 1px solid #ccc;
-  border-radius: 8px;
-  padding: 15px;
-  font-size: 14px;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-  max-width: 200px;
-  width: 100%;
-  aspect-ratio: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  box-sizing: border-box;
-}
-
-.info-box p {
-  margin: 5px 0;
-  color: #333;
-}
-
-.info-box button {
-  width: 100%;
-  padding: 8px;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 14px;
-  margin-top: 10px;
-}
-
-.add-btn {
-  background-color: #27ae60;
-  color: white;
-}
-
-.add-btn:hover {
-  background-color: #219a52;
-}
-
-.delete-btn {
-  background-color: #e74c3c;
-  color: white;
-}
-
-.delete-btn:hover {
-  background-color: #c0392b;
-}
-.boxes-container {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
-  gap: 15px;
-  position: relative;
-  width: 100%;
-  padding: 20px 0;
-  order: 0; /* This moves boxes above other content */
-}
-
-.form-group select:disabled {
-  opacity: 0.6;
-  cursor: not-allowed;
-}
-
-/* Modal */
-.modal-backdrop {
-  position: fixed;
-  inset: 0;
-  background: rgba(0,0,0,.5);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 9999;
-}
-
-.modal-panel {
-  width: min(900px, 92vw);
-  max-height: 85vh;
-  overflow: auto;
-  background: #fff;
-  border-radius: 14px;
-  padding: 18px 18px 12px;
-  box-shadow: 0 12px 40px rgba(0,0,0,.35);
-}
-
-.modal-panel h3 {
-  margin: 0 0 12px;
-  color: #1f3b8a;
-}
-
-.dog-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(180px, 1fr));
-  gap: 12px;
-  padding: 6px;
-}
-
-.dog-card {
-  border: 2px solid #e6e6e6;
-  background: #fdfdfd;
-  border-radius: 12px;
-  padding: 14px;
-  text-align: center;
-  box-shadow: 0 2px 8px rgba(0,0,0,.06);
-  cursor: pointer;
-}
-
-.dog-card:hover {
-  box-shadow: 0 4px 14px rgba(0,0,0,.12);
-}
-
-.dog-card.selected {
-  border-color: #ff8c00;
-  box-shadow: 0 0 0 3px rgba(255,140,0,.25);
-}
-
-.dog-card-name {
-  font-weight: 700;
-}
-
-.modal-actions {
-  display: flex;
-  justify-content: flex-end;
-  gap: 10px;
-  padding-top: 12px;
-}
-
-.modal-actions .ghost {
-  background: #fff;
-  border: 1px solid #ddd;
-}
-
-.modal-actions button {
-  padding: 8px 14px;
-  border-radius: 10px;
-  border: none;
-  background: #ff8c00;
-  color: #000000;
-  font-weight: 600;
-  cursor: pointer;
-}
-
-.modal-actions button:disabled {
-  opacity: .6;
-  cursor: not-allowed;
-}
-
-/* Pending-remove visual */
-.info-box.pending-remove {
-  opacity: 0.6;
-  background: #fff7f7;
-  border: 1px solid #f8c4c4;
-}
-
-.delete-btn.undo {
-  background: #6b7280; /* gray */
-}
