+++ frontend/src/routes/DashboardRoutes.tsx
@@ -12,7 +12,9 @@ const AdoptionPages = Loadable(lazy(() => import("../pages/public/adoption/admin
 const HealthRecords = Loadable(lazy(() => import("../pages/public/HealthRecord/SearchPage"))); // Placeholder for the health records component
 const CreateVisit = Loadable(lazy(() => import("../pages/dashboard/visit/createVisit")));
 const UpdateVisit = Loadable(lazy(() => import("../pages/dashboard/visit/updateVisit")));
-// ==============================|| DASHBOARD ROUTING ||============================== //
+const ZoneCageManagementPage = Loadable(lazy(() => import("../pages/dashboard/ZoneCageManagement")));
+const Volunteer = Loadable(lazy(() => import("../pages/dashboard/Volunteer")));
+/// ==============================|| DASHBOARD ROUTING ||============================== //
 export const dashboardRoutes: RouteObject = {
   path: "dashboard",
   element: <DashboardLayout />,
@@ -69,6 +71,14 @@ export const dashboardRoutes: RouteObject = {
     {
       path: "health-record/record/:recordId",
       element: <SingleDetailPageHealth />
+    },
+    {
+      path: "zone-cage-management",
+      element: <ZoneCageManagementPage />
+    },
+    {
+      path: "volunteer",
+      element: <Volunteer />
     }
   ],
 };
