+++ frontend/src/routes/ZCmanagementRoutes.tsx
@@ -3,7 +3,7 @@ import { lazy } from 'react';
 
 import Loadable from "../components/third-patry/Loadable";
 import FirstPage from "../pages/public/firstpage";
-const ZoneCageManagementPage = Loadable(lazy(() => import("../pages/public/ZoneCageManagement")));
+const ZoneCageManagementPage = Loadable(lazy(() => import("../pages/dashboard/ZoneCageManagement")));
 
 const ZCmanagementRoutes = (): RouteObject => {
   return {
